Output changes
I changed the q file so that we just have to set the usr and the file
will automatically find the right compilation command and paths.
Additionally, I added a unique ID for each tree to be able to identify
them across years and a second output file with values for each
individual tree.